remove useless things

+++ src/main/java/com/Nxer/TwistSpaceTechnology/system/ItemCooldown/IItemHasCooldown.java
@@ -3,5 +3,5 @@ package com.Nxer.TwistSpaceTechnology.system.ItemCooldown;
 public interface IItemHasCooldown {
 
     /** Returns the number of ticks of the cooldown */
-    int getCooldown();
+    long getCooldown();
 }
